server/requirements.txt
fastapi==0.115.9
openai==1.65.2
pydantic==2.10.6
python-dotenv==1.0.1
starlette==0.45.3
uvicorn==0.34.0